For #280: starting to reorganise code to allow using BLAST tabular - as opposed to only XMLs.
