+++ .gitattributes
@@ -1,0 +1,2 @@
+./sample_data/uniprot_sprot_plants.fasta.gz filter=lfs diff=lfs merge=lfs -text
+./sample_data/chr5.fas.gz filter=lfs diff=lfs merge=lfs -text
